LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/relations/css-styling/lengths-2-expected.txt


FAIL Units assert_equals: ex expected 500 but got 499.984375
FAIL Trimming of space assert_equals: ex expected 500 but got 499.984375
PASS Non-negative numbers
PASS Non-positive numbers
FAIL Legacy namedspaces assert_equals: veryverythinmathspace expected 0 but got 0.546875
FAIL Legacy numbers assert_equals: Unitless expected 30 but got 150
